backend/requirements.txt
asgiref==3.8.1
black==24.2.0
blinker==1.7.0
click==8.1.7
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
Flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.17.0
